Fully functioning up till ThreeNode
This marks a significant advance in the TwoThree tree, inasmuch as nodes
now morph cleanly between each other and we are bumping right up against
the breakthrough of ThreeNodes. After that our tree will have...
branches!

The big thing I learned in this commit was the ($SELF is rw : ... )
signature syntax. I must have seen it before, but now it makes sense :)
This gives you the container of 'self', which is what you need if you
want to, for instance, replace its contents with something else.

This is not likely to be a very nice thing to do for external users, so
the API is being thought of only in terms of what Tree will provide. The
rest will be lexical classes which remain internal to Tree.