Factored out the BUILD logic of TwoNode
We are already fully on the path of pushing the coercion cases to be
checked by multi signatures, this patch starts moving to align the
construction cases as well.

It's very pleasant to be able to 'domain refactor' this kind of thing on
the way through writing a library. Cheers to jnthn++ for cluing me onto
that term.