Once again re-aligning after a misunderstanding
In fact nodes should be Leafs until the point where they would become
TwoNode's. This is in contrast to promoting them while having their
Leafs be empty. So, a TwoNode should always have a $!d1 and nodes l/r
with at least one $!d1 as well.

With any luck, this is my last cockup :)